convert datatest tests cases to a custom test harness with libtest_mimic
