--- .cargo/config.toml
@@ -1,2 +1,0 @@
-[env]
-RUSTC_BOOTSTRAP = "1"
