+++ crates/wesl-test/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wgsl-test"
+name = "wesl-test"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
@@ -11,8 +11,13 @@ rust-version.workspace = true
 serde = { version = "1.0.210", features = ["derive"] }
 wgsl-parse = { workspace = true }
 wesl = { workspace = true, features = ["eval"] }
-datatest = "0.8.0"
 regex = "1.11.1"
+libtest-mimic = "0.8.1"
+serde_json = "1.0.140"
 
 [lints]
 workspace = true
+
+[[test]]
+name = "testsuite"
+harness = false
